requirements.txt
flask==2.0.1
opencv-python==4.5.3.56
mediapipe==0.8.9.1
numpy==1.21.2
Pillow==8.3.2
gunicorn==20.1.0